Updated arguments parsing.
