+++ segmlib/__init__.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+
+root_path = Path(__file__).resolve().parents[1]
+weights_path = root_path / 'weights'
+embeddings_path = root_path / 'embeddings'
